*** empty log message ***

+++ javascope/ASCIIDataProvider.java
@@ -235,12 +235,12 @@ class ASCIIDataProvider implements DataProvider
         return new SimpleWaveData(in_y, in_x);
     }
 
-    public WaveData GetResampledWaveData(String in, float start, float end, int n_points)
+    public WaveData GetResampledWaveData(String in, double start, double end, int n_points)
     {
         return null;
     }
 
-    public WaveData GetResampledWaveData(String in_y, String in_x, float start, float end, int n_points)
+    public WaveData GetResampledWaveData(String in_y, String in_x, double start, double end, int n_points)
     {
         return null;
     }
@@ -275,11 +275,11 @@ class ASCIIDataProvider implements DataProvider
         return new String(in);
     }
 
-    public float GetFloat(String in)
+    public double GetFloat(String in)
     {
         error = null;
-        Float f = new Float(in);
-        return f.floatValue();
+        Double f = new Double(in);
+        return f.doubleValue();
     }
